Improve database functions
Functions in database.py handle errors in database reads and writes by
logging the error to stdout and returning a boolean or int depending on
the function. Each of the functions establishes its own connection to
database so the caller does not need to worry about that. If the connection
can't be established the app terminates. Note GET by ID and POST are not
working at the time of this commit. They return a 'not supported' message.